+++ lib/features/sks_people_live/data/api_client/sks_client.dart
@@ -2,7 +2,7 @@ import "package:dio/dio.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
 
-import "../../../config/env.dart";
+import "../../../../config/env.dart";
 
 part "sks_client.g.dart";
